Added the bar chart to display runtime data.

+++ app/src/main/java/RunTimeComplexity/SetRunTime.java
@@ -6,7 +6,7 @@ import java.util.Set;
 public class SetRunTime {
 	private static final Random RAND = new Random();
 	
-	public static long searchTime(Set<Integer> set, int n) {
+	public static float searchTime(Set<Integer> set, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			set.contains(RAND.nextInt(n));
@@ -14,7 +14,7 @@ public class SetRunTime {
 		return Watch.getTime();
 	}
 	
-	public static long insertionTime(Set<Integer> set, int n) {
+	public static float insertionTime(Set<Integer> set, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			set.add(RAND.nextInt(n));
@@ -22,7 +22,7 @@ public class SetRunTime {
 		return Watch.getTime();
 	}
 	
-	public static long deletionTime(Set<Integer> set, int n) {
+	public static float deletionTime(Set<Integer> set, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			set.remove(RAND.nextInt(n));
